// src/types/Product.ts
export type Product = {
    id: number;
    name: string;
    type: string;
    calories: number;
    protein: number;
    carbs: number;
    sugar: number;
    fat: number;
    saturated_fat: number;
};

export type Meal = {
    id: number;
    name: string;
    type: string;
    calories: number;
    protein: number;
    carbs: number;
    sugar: number;
    fat: number;
    saturated_fat: number;
};